third_party/blink/web_tests/external/wpt/webrtc/RTCPeerConnection-setRemoteDescription-offer_interop-2026-expected.txt

This is a testharness.js-based test.
[FAIL] setRemoteDescription(offer) from have-local-offer fires signalingstatechange twice
  assert_equals: expected null but got object "[object RTCSessionDescription]"
[FAIL] setRemoteDescription(offer) in have-local-offer should update internal state with a queued task, in the right order
  assert_equals: pendingRemoteDescription should be updated before the signalingstatechange event expected null but got object "[object RTCSessionDescription]"
Harness: the test ran to completion.
